+++ cloudpilot-ionic/src/app/service/canvas-display.service.ts
@@ -160,6 +160,7 @@ const IMAGE_SILKSCREEN_PILOT = prepareImage('SILKSCREEN_PILOT');
 const IMAGE_SILKSCREEN_I705 = prepareImage('SILKSCREEN_I705');
 const IMAGE_SILKSCREEN_PEG_S300 = prepareImage('SILKSCREEN_PEG_S300');
 const IMAGE_SILKSCREEN_PEG_T415 = prepareImage('SILKSCREEN_PEG_T415');
+const IMAGE_SILKSCREEN_PEG_N610 = prepareImage('SILKSCREEN_PEG_N610');
 
 const IMAGE_BUTTONS_V = prepareImage('HARD_BUTTONS_PALM_V');
 const IMAGE_BUTTONS_M515 = prepareImage('HARD_BUTTONS_M515');
@@ -177,6 +178,7 @@ const IMAGE_BUTTONS_HANDERA330 = prepareImage('HARD_BUTTONS_HANDERA330');
 const IMAGE_HARD_BUTTONS_PEG_S300 = prepareImage('HARD_BUTTONS_PEG_S300');
 const IMAGE_HARD_BUTTONS_PEG_S320 = prepareImage('HARD_BUTTONS_PEG_S320');
 const IMAGE_HARD_BUTTONS_PEG_T415 = prepareImage('HARD_BUTTONS_PEG_T415');
+const IMAGE_HARD_BUTTONS_PEG_N610 = prepareImage('HARD_BUTTONS_PEG_N610');
 
 @Injectable({
     providedIn: 'root',
@@ -496,6 +498,9 @@ export class CanvasDisplayService {
 
     private silkscreenImage(): PrerenderedImage {
         switch (this.session?.device) {
+            case DeviceId.pegtN600c:
+                return IMAGE_SILKSCREEN_PEG_N610;
+
             case DeviceId.pegS320:
             case DeviceId.pegS300:
                 return IMAGE_SILKSCREEN_PEG_S300;
@@ -547,6 +552,9 @@ export class CanvasDisplayService {
 
     private buttonsImage(): PrerenderedImage {
         switch (this.session?.device) {
+            case DeviceId.pegtN600c:
+                return IMAGE_HARD_BUTTONS_PEG_N610;
+
             case DeviceId.pegS300:
                 return IMAGE_HARD_BUTTONS_PEG_S300;